[theme] move .*Color tests to color_test.go
